Add logic to connect to WiFi and display status on LED matrix
If there are WiFi credentials from a previous setup, the program attempts to
connect to that WiFi network.

In case it fails to establish a connection, then it launches smartconfig to
receive the WiFi credentials. The user gives them through the ESPTouch android
app.

Once the device is connected to the WiFi network, the name of the project is
displayed on the LED matrix. This is a temporary feature since later this will
be updated.