Implement logic to establish a connection to a MQTT broker
If after several attempts to connect to a MQTT broker is not possible to establish
the connection, then display the status on the LED matrix.

For the moment the broker's IP address is hardcoded but this will be changed in
the future so the user can give the IP address through a PC or Android application.